backend/requirements.txt
requests~=2.32.3
psycopg2-binary~=2.9.10
fastapi~=0.115.3
uvicorn~=0.32.0